+++ RoSA_Monitor/launchermanager.cpp
@@ -102,6 +102,14 @@ void LauncherManager::StopLauncher(LauncherType type){
          case LIDAR:
              auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 launch urg_node2 urg_node2.launch.py";
              break;
+         case SLAM:
+             auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && cd maps && ros2 launch rosa_description amcl_launch.py use_sim_time:=false slam:=True "
+                                                                "slam_params:=" + workspace_dir->path() + "/src/rosa_description_Pablo/config/mapper_params_online_async.yaml "
+                                                                "params_file:=" + workspace_dir->path() + "/src/rosa_description_Pablo/config/nav2_params.yaml";
+             break;
+         case NAVIGATION:
+             auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 launch rosa_description amcl_launch.py use_sim_time:=false map:=" + mapLoaded->fileName();
+             break;
          case GAZEBO_SIM:
              auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 launch rosa_description rosa_gazebo_launch.py use_sim_time:=true";
              break;
@@ -110,7 +118,7 @@ void LauncherManager::StopLauncher(LauncherType type){
                                                                 "slam_params:=" + workspace_dir->path() + "/src/rosa_description_Pablo/config/mapper_params_online_async.yaml "
                                                                 "params_file:=" + workspace_dir->path() + "/src/rosa_description_Pablo/config/nav2_params.yaml";
              break;
-         case NAVIGATION:
+         case NAVIGATION_SIM:
              auxCommand = "cd && cd " + workspace_dir->path() + " && source install/setup.bash && ros2 launch rosa_description amcl_launch.py use_sim_time:=true map:=" + mapLoaded->fileName();
              break;
          default: